VC.jl API Reference

VC.jl API Reference

VC

VC.ImageTensorConversion

External

Index

VC

VC.expand_to_4x4 — Method

expand_to_4x4(matrix::AbstractMatrix{T}) where {T}

Expands a 3x3 matrix to a 4x4 matrix. The element at [4,4] will be a 1.

Arguments

matrix: The 3x3 matrix to expand.

VC.gpu — Method

gpu(arr::AbstractArray{T})::AbstractArray{T} where {T}

The available GPU backends are:

CUDA (Nvidia)

AMDGPU (AMD)

oneAPI (Intel)

Metal (Apple)

The extension will be loaded when the GPU backend is loaded. E.g.

julia> using CUDA

will load the CUDA extension.

If no backend is loaded this function does nothing.

VC.imread — Method

imread([T = Float32], path::String)::AbstractArray{T, 3} where {T <: AbstractFloat}

Loads the given image as a HxWxC tensor.

See also: tensor, ImageTensorConversion.image, imshow

Arguments

T: Type of the elements of the resulting tensor. Defaults to Float32.

path: Path to the image to load.

VC.imshow — Method

imshow(img::AbstractArray{T, 2}; show=true, save_to=Nothing) where {T <: Colorant}

Shows and/or saves the given image. Waits until the image is closed.

See also: imread, ImageTensorConversion.image, tensor.

Arguments

img: The image to display and/or save.

show: Determines wheter the image should be displayed.

save_to: Location to save the image to. Defaults to Nothing which will not save the image.

VC.linspace — Method

linspace([T = Float32], start, finish, steps::Integer)::AbstractArray{T, 1} where {T <: AbstractFloat}

Returns a vector of elements of type T from start to finish in steps steps.

Arguments

T: Type of the elements in the resulting vector. Defaults to Float32.

start: The first element in the resulting vector.

finish: The last element in the resulting vector.

steps: The number of steps to split the range into.

Example

julia> linspace(0, 10, 11)
11-element Vector{Float32}:
  0.0
  1.0
  2.0
  3.0
  4.0
  5.0
  6.0
  7.0
  8.0
  9.0
 10.0

VC.makegrid — Method

makegrid(
    images::AbstractVecOrMat{<:AbstractArray{T, 3}},
    dims::NTuple{2, <:Integer}
    )::AbstractArray{T, 3} where {T <: AbstractFloat}

Aranges the images in a grid.

Arguments

images: The images to arange in a grid.

dims: The dimensions of the grid.

Example

julia> images = [ rand(64, 64, 3) for _ in 1:16 ]; # 16 images of random noise

julia> grid = makegrid(images, (4, 4));

julia> size(grid)
(256, 256, 3)

VC.orthogonalize — Method

orthogonalize(M::AbstractMatrix{T})::AbstractMatrix{T} where {T <: AbstractFloat}

Computes the orthogonal basis to the two column vectors defined in the matrix M as a matrix.

Arguments

M: The matrix containing the two vectors to compute the basis for.

VC.rotation_from_axis_angle — Method

rotation_from_axis_angle(axis::AbstractVector{T}, θ::T) where {T <: AbstractFloat}

Computes a 3x3 rotation matrix from a given axis and angle.

Arguments

axis: The axis to rotate around.

θ: The angle to rotate by.

VC.rotation_from_quaternion — Method

rotation_from_quaternion(quaternion::AbstractVector{T}) where { T <: AbstractFloat }

Computes a 3x3 rotation matrix from a given quaternion.

Arguments

quaternion: The quaternion encoding the rotation.

VC.row_mul — Method

row_mul(M::AbstractMatrix{T}, vs::AbstractMatrix{T})::AbstractMatrix{T} where {T <: Real}

This function multiplies the matrix M onto every row of the matrix vs.

Arguments

M: The matrix to multiply onto the rows.

vs: The vectors to mulltiply M onto stored in the rows of a matrix.

VC.show_by_default! — Method

show_by_default!(value::Bool)

Sets wether or not to show the image when calling imshow by default.

Arguments

value: New value for the flag.

VC.ImageTensorConversion

VC.ImageTensorConversion.image — Method

image(t::AbstractArray{T, 3})::AbstractArray{U, 2} where {T <: AbstractFloat, U <: Colorant}

Converts a HxWxC tensor to a displayable image. tensor is the inverse to this function.

See also: tensor, VC.imshow, VC.imread.

Arguments

t: HxWxC tensor containing the color data of an image

Example

julia> t = rand(144, 256, 3); # image data of an RGB 256x144 image

julia> img = t |> image; # 144x256 matrix containing RGB values

julia> size(img)
(144, 256)

julia> typeof(img[1])
RGB{Float64}

VC.ImageTensorConversion.tensor — Method

tensor([T = Float32], img::AbstractArray{RGB, 2})::AbstractArray{T, 3}

Converts an image to a HxWxC tensor. image is the inverse to this function.

See also: image, VC.imshow, VC.imread.

Arguments

T: Type of the elements of the resulting tensor. Defaults to Float32.

img: HxW matrix containing the RGB values of the image.

Example

julia> img = rand(RGB, 144, 256);

julia> size(img)
(144, 256)

julia> t = img |> tensor;

julia> size(t)
(144, 256, 3)

julia> typeof(t[1])
Float32

julia> t = tensor(Float64, img);

julia> typeof(t[1])
Float64

External

You can find the documentation of the most important reexported packages here:

Linear Algebra: LinearAlgebra

Machine Learning: Lux

Plotting: Makie

Optimization: Optimisers

Automatic Differentiation: Zygote

The full reexport looks like this:

using Reexport
@reexport using FileIO, ImageIO, MeshIO, CairoMakie
@reexport using LinearAlgebra, Statistics, Printf, Random, ProgressMeter
@reexport using StaticArrays, KernelAbstractions
@reexport using Lux, Zygote, Optimisers
@reexport import ColorTypes

Index

VC.ImageTensorConversion.image

VC.ImageTensorConversion.tensor

VC.expand_to_4x4

VC.gpu

VC.imread

VC.imshow

VC.linspace

VC.makegrid

VC.orthogonalize

VC.rotation_from_axis_angle

VC.rotation_from_quaternion

VC.row_mul

VC.show_by_default!